Add files via upload
overview:
the us healthcare department has shared the hospital data for 100 hospitals which includes data like patient information, their admission information and the insurance information. the top executives from the department have a few questions on the data and would like you to analyses and present your findings to them. the analysis should derive meaning actionable insights and the presentation should be clear, complete and concise.